Rename libpfasst_swe_sphere -> libpfasst_swe_sphere_mlsdc
* update compile test accordingly
* update convergence test accordingly
* add warning that program is not ready for use
* modify convergence test
