+++ src/programs/libpfasst_swe_sphere_mlsdc.cpp
@@ -1,6 +1,6 @@
 /*
  * Author: Francois Hamon & Martin Schreiber <SchreiberX@gmail.com>
- * MULE_COMPILE_FILES_AND_DIRS: src/programs/libpfasst_swe_sphere
+ * MULE_COMPILE_FILES_AND_DIRS: src/programs/libpfasst_swe_sphere_mlsdc
  * MULE_COMPILE_FILES_AND_DIRS: src/programs/swe_sphere_timeintegrators/
  * MULE_COMPILE_FILES_AND_DIRS: src/programs/swe_sphere_benchmarks/
  * MULE_SCONS_OPTIONS: --sphere-spectral-space=enable
@@ -13,7 +13,7 @@
 #include <sweet/sphere/SphereOperators_SphereData.hpp>
 #include <sweet/SimulationVariables.hpp>
 #include "libpfasst_interface/LevelSingleton.hpp"
-#include "libpfasst_swe_sphere/SphereDataCtx.hpp"
+#include "libpfasst_swe_sphere_mlsdc/SphereDataCtx.hpp"
 #include "swe_sphere_benchmarks/BenchmarksSphereSWE.hpp"
 #include <sweet/SWEETError.hpp>
 
@@ -46,6 +46,10 @@ void fmain (SphereDataCtx* pd_ctx,
 int main(int i_argc, char *i_argv[])
 {
 	MPI_Init(&i_argc, &i_argv);
+
+	std::cout << "+++++++++++++++++++++++++++++++++++++++++++" << std::endl;
+	std::cout << "WARNING: This program is NOT ready for use!" << std::endl;
+	std::cout << "+++++++++++++++++++++++++++++++++++++++++++" << std::endl;
 
 	// input parameter names (specific ones for this program)
 	const char *bogus_var_names[] = {
